Add test for directory imports
Was running a binary from an old build that had a bug around
this functionality. App would crash if one attempted to import
an entire directory. Wasted time writing a unit test that
exposes the bug, only to find out that the bug was fixed
(well I at least raised the test coverage).